+++ test/utils/DeployOptimized.sol
@@ -4,49 +4,33 @@ pragma solidity >=0.8.19 <0.9.0;
 import { IAllowanceTransfer } from "@uniswap/permit2/interfaces/IAllowanceTransfer.sol";
 import { StdCheats } from "forge-std/StdCheats.sol";
 
-import { ISablierV2Archive } from "../../src/interfaces/ISablierV2Archive.sol";
 import { ISablierV2Batch } from "../../src/interfaces/ISablierV2Batch.sol";
 import { ISablierV2MerkleStreamerFactory } from "../../src/interfaces/ISablierV2MerkleStreamerFactory.sol";
-import { ISablierV2ProxyPlugin } from "../../src/interfaces/ISablierV2ProxyPlugin.sol";
 import { SablierV2ProxyTargetApprove } from "../../src/SablierV2ProxyTargetApprove.sol";
 import { SablierV2ProxyTargetPermit2 } from "../../src/SablierV2ProxyTargetPermit2.sol";
 import { SablierV2ProxyTargetPush } from "../../src/SablierV2ProxyTargetPush.sol";
 
 abstract contract DeployOptimized is StdCheats {
-    /// @dev Deploys {SablierV2Archive} from a optimized source compiled with `--via-ir`.
-    function deployOptimizedArchive(address initialAdmin) internal returns (ISablierV2Archive) {
-        return ISablierV2Archive(
-            deployCode("out-optimized/SablierV2Archive.sol/SablierV2Archive.json", abi.encode(initialAdmin))
-        );
-    }
-
-    /// @dev Deploys {SablierV2Batch} from a optimized source compiled with `--via-ir`.
+    /// @dev Deploys {SablierV2Batch} from an optimized source compiled with `--via-ir`.
     function deployOptimizedBatch() internal returns (ISablierV2Batch) {
         return ISablierV2Batch(deployCode("out-optimized/SablierV2Batch.sol/SablierV2Batch.json"));
     }
 
-    /// @dev Deploys {SablierV2MerkleStreamerFactory} from a optimized source compiled with `--via-ir`.
+    /// @dev Deploys {SablierV2MerkleStreamerFactory} from an optimized source compiled with `--via-ir`.
     function deployOptimizedMerkleStreamerFactory() internal returns (ISablierV2MerkleStreamerFactory) {
         return ISablierV2MerkleStreamerFactory(
             deployCode("out-optimized/SablierV2MerkleStreamerFactory.sol/SablierV2MerkleStreamerFactory.json")
         );
     }
 
-    /// @dev Deploys {SablierV2ProxyPlugin} from a optimized source compiled with `--via-ir`.
-    function deployOptimizedProxyPlugin(ISablierV2Archive archive_) internal returns (ISablierV2ProxyPlugin) {
-        return ISablierV2ProxyPlugin(
-            deployCode("out-optimized/SablierV2ProxyPlugin.sol/SablierV2ProxyPlugin.json", abi.encode(archive_))
-        );
-    }
-
-    /// @dev Deploys {SablierV2ProxyTargetApprove} from a optimized source compiled with `--via-ir`.
+    /// @dev Deploys {SablierV2ProxyTargetApprove} from an optimized source compiled with `--via-ir`.
     function deployOptimizedProxyTargetApprove() internal returns (SablierV2ProxyTargetApprove) {
         return SablierV2ProxyTargetApprove(
             deployCode("out-optimized/SablierV2ProxyTargetApprove.sol/SablierV2ProxyTargetApprove.json")
         );
     }
 
-    /// @dev Deploys {SablierV2ProxyTargetPermit2} from a optimized source compiled with `--via-ir`.
+    /// @dev Deploys {SablierV2ProxyTargetPermit2} from an optimized source compiled with `--via-ir`.
     function deployOptimizedProxyTargetPermit2(IAllowanceTransfer permit2_)
         internal
         returns (SablierV2ProxyTargetPermit2)
@@ -58,7 +42,7 @@ abstract contract DeployOptimized is StdCheats {
         );
     }
 
-    /// @dev Deploys {SablierV2ProxyTargetPush} from a optimized source compiled with `--via-ir`.
+    /// @dev Deploys {SablierV2ProxyTargetPush} from an optimized source compiled with `--via-ir`.
     function deployOptimizedProxyTargetPush() internal returns (SablierV2ProxyTargetPush) {
         return SablierV2ProxyTargetPush(
             deployCode("out-optimized/SablierV2ProxyTargetPush.sol/SablierV2ProxyTargetPush.json")
@@ -67,34 +51,24 @@ abstract contract DeployOptimized is StdCheats {
 
     /// @notice Deploys all V2 Periphery contracts from a optimized source in the following order:
     ///
-    /// 1. {SablierV2Archive}
-    /// 2. {SablierV2Batch}
-    /// 3. {SablierV2MerkleStreamerFactory}
-    /// 4. {SablierV2ProxyPlugin}
-    /// 5. {SablierV2ProxyTargetApprove}
-    /// 6. {SablierV2ProxyTargetPermit2}
-    /// 7. {SablierV2ProxyTargetPush}
-    function deployOptimizedPeriphery(
-        address initialAdmin,
-        IAllowanceTransfer permit2_
-    )
+    /// 1. {SablierV2Batch}
+    /// 2. {SablierV2MerkleStreamerFactory}
+    /// 3. {SablierV2ProxyTargetApprove}
+    /// 4. {SablierV2ProxyTargetPermit2}
+    /// 5. {SablierV2ProxyTargetPush}
+    function deployOptimizedPeriphery(IAllowanceTransfer permit2_)
         internal
         returns (
-            ISablierV2Archive,
             ISablierV2Batch,
             ISablierV2MerkleStreamerFactory,
-            ISablierV2ProxyPlugin,
             SablierV2ProxyTargetApprove,
             SablierV2ProxyTargetPermit2,
             SablierV2ProxyTargetPush
         )
     {
-        ISablierV2Archive archive_ = deployOptimizedArchive(initialAdmin);
         return (
-            archive_,
             deployOptimizedBatch(),
             deployOptimizedMerkleStreamerFactory(),
-            deployOptimizedProxyPlugin(archive_),
             deployOptimizedProxyTargetApprove(),
             deployOptimizedProxyTargetPermit2(permit2_),
             deployOptimizedProxyTargetPush()